🚚 Move Spotify now playing lib to spotify/now-playing
